feat(host): separate scenarios and run sequentially
