+++ makefile
@@ -13,7 +13,7 @@ test:
 	TEST=1 RUN=1 ./script/build.sh $(TARGET)
 
 integration:
-	CONFIG_PATH="config/config.json" cargo test -F integration proof_request
+	CONFIG_PATH="config/config.json" cargo test -F integration run_scenarios_sequentially
 
 fmt:
 	@cargo fmt --all --check
